+++ lib/src/rules/prule_cpf.dart
@@ -17,19 +17,22 @@ class PRuleCpf implements Rule {
     "12345678909"
   ];
 
-  String? _value;
-  String message;
+  final String? _value;
+  final String message;
 
+  /// Check if the value is a valid CPF
+  /// [value] is the value to check
+  /// [message] is the error message
   PRuleCpf(this._value, {this.message = "Invalid CPF"});
 
   @override
   String? validate() {
-    if (this._value == null || this._value?.trim().length == 0) {
+    if (_value == null || _value?.trim().length == 0) {
       return null;
-    } else if (this.isValid((this._value))) {
+    } else if (isValid((_value))) {
       return null;
     } else {
-      return this.message;
+      return message;
     }
   }