+++ lib/src/rules/prule_max_numeric.dart
@@ -1,21 +1,23 @@
 import 'package:pvalidator/src/rules/rule.dart';
 
 class PRuleMaxNumeric implements Rule {
-  String? _value;
-  double max;
-  String message;
+  final String? _value;
+  final double max;
+  final String message;
 
+  /// Check if the value less than or equal to the max value
+  /// [value] is the value to check
+  /// [max] is the maximum value
+  /// [message] is the error message
   PRuleMaxNumeric(this._value, this.max, {this.message = "Maximum value is"});
 
   @override
   String? validate() {
-    double? doubleValue = double.tryParse(this._value ?? '');
-    if (this._value == null ||
-        this._value?.trim().length == 0 ||
-        doubleValue == null) {
+    double? doubleValue = double.tryParse(_value ?? '');
+    if (_value == null || _value?.trim().length == 0 || doubleValue == null) {
       return null;
-    } else if (doubleValue > this.max) {
-      return "${this.message} ${this.max.toString()}";
+    } else if (doubleValue > max) {
+      return "$message ${max.toString()}";
     } else {
       return null;
     }