+++ lib/src/rules/prule_required.dart
@@ -1,15 +1,18 @@
 import 'package:pvalidator/src/rules/rule.dart';
 
 class PRuleRequired implements Rule {
-  String? _value;
-  String message;
+  final String? _value;
+  final String message;
 
+  /// Check if the value is present
+  /// [value] is the value to check
+  /// [message] is the error message
   PRuleRequired(this._value, {this.message = "Required"});
 
   @override
   String? validate() {
-    if (this._value == null || this._value?.trim().length == 0) {
-      return this.message;
+    if (_value == null || _value?.trim().length == 0) {
+      return message;
     } else {
       return null;
     }