+++ lib/src/rules/prule_requiredif.dart
@@ -1,19 +1,23 @@
 import 'package:pvalidator/src/rules/rule.dart';
 
 class PRuleRequiredIf implements Rule {
-  String? _value;
-  bool _ifRequired;
-  String message;
+  final String? _value;
+  final bool _ifRequired;
+  final String message;
 
+  /// Check if the value is present if the condition is true
+  /// [value] is the value to check
+  /// [ifRequired] is the condition to check
+  /// [message] is the error message
   PRuleRequiredIf(this._value, this._ifRequired, {this.message = "Required"});
 
   @override
   String? validate() {
-    if (!this._ifRequired) {
+    if (!_ifRequired) {
       return null;
     }
-    if (this._value == null || this._value?.trim().length == 0) {
-      return this.message;
+    if (_value == null || _value?.trim().isEmpty == true) {
+      return message;
     } else {
       return null;
     }