+++ lib/src/rules/prule_size_string.dart
@@ -1,18 +1,22 @@
 import 'package:pvalidator/src/rules/rule.dart';
 
 class PRuleSizeString implements Rule {
-  String? _value;
-  int size;
-  String message;
+  final String? _value;
+  final int size;
+  final String message;
 
+  /// Check if the value size is equal to the size value
+  /// [value] is the value to check
+  /// [size] is the size of value
+  /// [message] is the error message
   PRuleSizeString(this._value, this.size, {this.message = "Size must be"});
 
   @override
   String? validate() {
-    if (this._value == null || this._value?.trim().length == 0) {
+    if (_value == null || _value?.trim().length == 0) {
       return null;
-    } else if (this._value!.length != this.size) {
-      return "${this.message} ${this.size}";
+    } else if (_value!.length != size) {
+      return "$message $size";
     } else {
       return null;
     }